+++ src/share/classes/com/sun/javatest/regtest/agent/MainWrapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -134,6 +134,10 @@ public class MainWrapper
 
                 // RUN JAVA PROGRAM
                 Class<?> c = Class.forName(className, false, cl);
+                if (MainMethodHelper.isCompactSourceFileAndInstanceMainMethodSupported()) {
+                    MainMethodHelper.executeMainClass(c, args);
+                    return;
+                }
                 Method mainMethod = c.getMethod("main", String[].class);
                 mainMethod.invoke(null, (Object) args);